Add idea of comesWith units to UnitTemplates.
These are units that cannot be bought separately but come with the purchase of another unit.
The minion units have a new flag set - notForPurchase - and are not displayed for purchase in the team edit template.
The master units have a new ManyToMany field set containing the 1 to N unit templates that come with them.
These units are automatically added to the team after the purchase of the master. Only the master unit is checked for exceeding any team limits.
Minion unit names are shown in [ brackets ] below their master unit when listing units for purchase.